refactor(users): relocate Role enum to domain-pure module
Move `Role` from `app/users/models.py` (which transitively loads
SQLAlchemy) to a new domain-pure module `app/users/domain/value_objects.py`
so that `users/domain/` no longer depends on the ORM module. Resolves the
last `domain/` deviation note in `app/users/domain/__init__.py`.

`ServerType` (the other deviation tracked in #235) is deferred to land
alongside #228 (servers domain repository introduction), per the issue's
recommended approach B.

Changes:
- New `app/users/domain/value_objects.py` defining `Role` (stdlib only)
- `app/users/models.py` imports `Role` from the new module for its
  SQLAlchemy `Enum` column reference
- All ~45 callsites (app/ + tests/) updated to import `Role` from
  `app.users.domain.value_objects`
- Pilot-deviation docstring removed from `app/users/domain/__init__.py`
  and `app/users/domain/entities.py`

Resolves a portion of #235 (Role half). `ServerType` half tracked under #228.
