+++ app/files/router.py
@@ -31,7 +31,8 @@ from app.services.authorization_service import authorization_service
 from app.services.file_history_service import file_history_service
 from app.services.file_management_service import file_management_service
 from app.types import FileType
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 router = APIRouter()
