+++ app/services/server_service.py
@@ -9,7 +9,8 @@ from sqlalchemy.orm import Session
 
 from app.servers.models import Server, ServerStatus, ServerType
 from app.services.minecraft_server import minecraft_server_manager
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 logger = logging.getLogger(__name__)
