+++ app/services/visibility_service.py
@@ -17,7 +17,8 @@ from app.core.visibility import (
     ResourceVisibility,
     VisibilityType,
 )
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 logger = logging.getLogger(__name__)
