+++ tests/unit/models/test_backup_schedule_models.py
@@ -4,7 +4,8 @@ from sqlalchemy.orm import Session
 
 from app.backups.models import BackupSchedule, BackupScheduleLog, ScheduleAction
 from app.servers.models import Server
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 
 class TestBackupScheduleModel:
