+++ tests/unit/services/test_authorization_service_phase2.py
@@ -16,7 +16,8 @@ from app.core.visibility import (
     VisibilityType,
 )
 from app.services.authorization_service import AuthorizationService
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 
 class TestAuthorizationServicePhase2Visibility:
