+++ tests/unit/services/test_backup_schedule_service.py
@@ -5,7 +5,8 @@ from sqlalchemy.orm import Session
 
 from app.backups.models import BackupSchedule, BackupScheduleLog, ScheduleAction
 from app.servers.models import Server
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 
 class TestBackupScheduleService:
