+++ tests/unit/services/test_phase1_compliance.py
@@ -17,7 +17,8 @@ from sqlalchemy.orm import Session
 
 from app.servers.models import Server, ServerStatus, ServerType
 from app.services.authorization_service import AuthorizationService
-from app.users.models import Role, User
+from app.users.domain.value_objects import Role
+from app.users.models import User
 
 
 class TestPhase1ComplianceBasicOperations:
